2007-10-16 (Tebasuna) + New RaWav.extension and AviSynth plugin RaWav.dll. Open files > 4 GB with uncompressed formats like: wav (WAVE_FORMAT_EXTENSIBLE and others), bwf, raw, au, aif, w64, RF64, or caf (see RaWav_readme)
2007-08-13 (Tebasuna)
+ New changes to support new enc_aacPlus.exe from Shon3i: select between Mp4Box and Mp4Mux. (CodingTechnologiesAAC.cs)
+ New FraunhoferMp3.extension to support free mp3sEncoder.exe from Fraunhofer (http://www.all4mp3.com/tools/sw_fhg_cl.html).
+ New wavSplit.extension and wavSplit.exe to replace wav2mono and wav2stereo
+ Only literals changes. (AssemblyInfo.cs, OggVorbisEncoder.cs)

2007-05-16 (thuongshoo)
+ Graphic changes and approximate ogg bitrate when VBR. (OggVorbisEncoder.cs, CodingTechnologiesAAC.cs)

2007-04-01 (Tebasuna)
+ Changes to support new enc_aacPlus.exe from Shon3i. (NeroDigitalEncoder.cs, CodingTechnologiesAAC.cs)

2007-03-31 (Chumbo)
+ updated TimeStretch rate control radio button labels per suggestions by tebasuna, alwa and madshi.
+ added hints/popup text to rate control radio buttons. I wound up using the AVISynth text for the hint text as it's straightforward.

2007-03-30 (Chumbo)
+ added rate control to the TimeStretch DSP module - you can select rate, pitch or both
+ updated the version to 0.1.10.* from 0.1.9.*. With all the changes we made we should've update the version sooner

2007-03-28 (Shon3i)
+ New enc_aacPlus.exe version to support enc_aacPlus.dll from WinAmp 5.33/5.34 and new MP4mux instead MP4Box.(enc_aacPlus.exe, MP4mux.exe)

2007-03-21 (Chumbo)
+ Destination/output dialog will now pre-fill the file name

2007-03-20 (Chumbo)
+ fixed a small UI bug to make sure when a job is finished, the item selected doesn't get deselected in the jobs list

2007-03-20 (Tebasuna)
+ problem fixed because AviSynth bug: IsAudioInt return always false (LameEncoder.cs, MusePack.extension, ffmpeg.extension)

2007-03-19 (Alwa)
+ fixed a problem with the "encoder" directory in connection with the WAVWriter output(MainForm.cs)

2007-03-18 (Chumbo)
+ added accelerators to buttons so keyboard can be quickly used to access functions
+ when either source/destination are blank, clicking Enqueue displays a message
+ "Keep output..." check box now works AFTER you've started a process
+ Move buttons in Queue tab are now enabled based on list selections (1 item selected only)
+ fixed unintended multiple selections in Jobs list. Now that we have a multi-select list, certain
processes were causing items to stay selected when they should not.

2007-03-18 (Alwa)
+ renamed the "ext" directory to "extensions" (MainForm.cs, BeHappy.csproj)
+ Encoder binaries can be placed in the "encoder" directory(handling like extensions directory)(MainForm.cs)

2007-03-18 (Tebasuna)
+ Without mod in code. Cancell warnings and copy files in compile time (MainForm.cs, Encoder.cs, compile.bat)

2007-03-16 (Chumbo)
+ reinstated code from 03-07 build that fell through the cracks during the incremental builds
+ fixed the "Keep Output..." checkbox functionality

2007-03-15 (Tebasuna)
+ Added 'Additional CLI arguments' to Aften capture (AftenEncoder.cs, MainForm.cs).
+ Optional 'pseudo-encoders' to split in mono or stereo wav's (wav2mono.*, Wav2stereo.*)
+ Optional optimized Downmix routines. Fast and without inversion in rear channel. (DownMix.extension).

2007-03-14 (Chumbo)
+ Just made sure that the EnsureVBRMP3Sync default is false per Tebasuna's earlier change

2007-03-14 (Tebasuna)
+ Adjust to best quality encode possible. The old method can be obtained with a final DSP Convert Sample to 16 bit int.
(LameEncoder.cs, OggVorbisEncoder.cs, CodingTechnologiesAAC.cs, aud-x.extension, ffmpeg.extension, flac.extension)
+ Added twolame.extension for twolame v0.38 (.mp2) http://www.rarewares.org/. For use instead ffmpeg.
+ Added musepack.extension for mppenc v1.16 (.mpc) http://www.musepack.net/

2007-03-13 (Chumbo)
+ modified Delete All button behavior to only remove items with status of Done, Error and Abort (MainForm.cs)
+ changed job list control to a multi-select control
+ modified Delete button behavior to remove the selected items except if item selected is in Processing state

2007-03-13 (Alwa)
+ fixed the preferMP4overM4A handling (NeroDigitalEncoder.cs).

2007-03-12 (Tebasuna)
+ Changes to allow any AviSynth output samples. Compatible with actual and alternative AvisynthWrapper.dll (Encoder.cs)
+ New ConvertSample.extension. Needed to test output samples and check encoders:
The old AvisynthWrapper.dll is equivalent to use the new one with a final DSP Convert Sample to 16 bit int.
Warning, some encoders need 16 bit int output (test pending).
+ Descriptive literal change: "Custom (actual_time / desired_time) x 100" instead "Custom transform" (TimeStretchDSP.cs)
+ EnsureMP3VBRSync unchecked by default, in my opinion not needed in BeHappy (MainForm.cs).

2007-03-12 (Chumbo)
+ Alternative AvisynthWrapper.dll without the fix ConvertAudioTo16bit output (AvisynthWrapper.cpp).
+ replaced deprecated strcpy calls with new strcpy_s method (AvisynthWrapper.cpp).

2007-03-11 (Alwa)
+ .config moved to the BeHappy.state file (unlovely, but it works) (NeroDigitalEncoder.cs).
+ CT Changes with multiplier 1000 (CodingTechnologiesAAC.cs).
+ extensions can now be stored either in the "ext" directory or like before (MainForm.cs, BeHappy.csproj).
+ DirectShowSource(... , video=false) (Extension.cs).
+ moved GuiPosition class to Configuration.cs (Configuration.cs, Job.cs).

2007-03-10 (Alwa)
+ made the Ensure MP3 VBR Sync checkbox work (MainForm.cs).
+ target filename will change now after loading an new source file (MainForm.cs).
+ aften encoder GUI bitrate trackbar improvement (AftenEncoder.cs).

2007-03-09 (Chumbo)
+ default window size is 800x400 now (MainForm.cs,Configuration.cs, Job.cs).
+ once you close it it'll save the size and position state

2007-03-07 (Chumbo)
+ added Delete All to quickly remove all the jobs except the one that is processing (MainForm.cs).
+ added check box to allow you to keep any output created if you decide to abort early or if an error occurs (MainForm.cs).

2007-03-05 (Tebasuna)
+ Aften CBR bitrate rounded to a valid value when is send to Aften (AftenEncoder.cs).
+ Aften Dinamic Range Compression instead BandWidth (AftenEncoder.cs).
+ Aften Read to End of File instead Selectively use 256-point MDCT (AftenEncoder.cs).

2007-03-03 (Chumbo)
+ Aften CBR max bitrate 640 (AftenEncoder.cs).

+++ BeHappy/Encoder.cs
@@ -20,7 +20,8 @@ namespace BeHappy
 			Notify,
 			Progress,
 			StdErr,
-			StdOut
+			StdOut,
+            KeepOutput
 		}
 
 		internal string Message = null;
@@ -56,6 +57,7 @@ namespace BeHappy
 		private string	m_encoder;
 		private string	m_commandLine;
 		private bool	m_sendHeader;
+        private bool m_bKeepOutput;
 
 		private long m_nSampleCount;
 		private long m_nSizeInBytes;
@@ -86,19 +88,28 @@ namespace BeHappy
 
 		private void audioStreamInfo(AviSynthClip x)
 		{
-			raiseEvent(new EncoderCallbackEventArgs(string.Format("Channels={0}, BitsPerSample={1}, SampleRate={2}Hz", x.ChannelsCount, x.BitsPerSample, x.AudioSampleRate)));
+// New for float output
+       			if(x.SampleType==AudioSampleType.FLOAT)
+       			{
+                 		raiseEvent(new EncoderCallbackEventArgs(string.Format("Channels={0}, BitsPerSample={1} float, SampleRate={2}Hz", x.ChannelsCount, x.BitsPerSample, x.AudioSampleRate)));
+       			}
+       			else
+       			{
+       	                	raiseEvent(new EncoderCallbackEventArgs(string.Format("Channels={0}, BitsPerSample={1} int, SampleRate={2}Hz", x.ChannelsCount, x.BitsPerSample, x.AudioSampleRate)));
+       			}
 		}
 
 
 
-		internal Encoder(string script, string output, string encoder, string commandLine, bool sendHeader)
+        internal Encoder(string script, string output, string encoder, string commandLine, bool sendHeader, bool bKeepOutput)
 		{
 			this.m_script = script;
 			this.m_output = output;
 			this.m_encoder = encoder;
 			this.m_commandLine = commandLine;
 			this.m_sendHeader = sendHeader;
-			m_process = null;
+            this.m_bKeepOutput = bKeepOutput;
+            m_process = null;
 			if(null!=m_encoder)
 				if(0==m_encoder.Length)
 					m_encoder=null;
@@ -106,12 +117,13 @@ namespace BeHappy
 
 		internal Encoder(
 			Job job):this(
-			job.AviSynthScript, 
-			job.TargetFile, 
-			job.EncoderExecutable, 
-			job.CommandLine, 
-			job.SendRiffHeader)
-		{
+			job.AviSynthScript,
+			job.TargetFile,
+			job.EncoderExecutable,
+			job.CommandLine,
+            job.SendRiffHeader,
+            job.bKeepOutput)
+        {
 		}
 
 		private void readStdOut()
@@ -220,8 +232,9 @@ namespace BeHappy
 									}
 
 									setProgress(100);
-									if(m_sendHeader && (x.SampleType==AudioSampleType.INT24 || x.SampleType==AudioSampleType.INT8))
-										target.WriteByte(0);
+// Next lines commented. New to allow output all samples. I can't understand for what write a extrabyte 0 at end of 8 and 24 int samples, cancelled.
+//									if(m_sendHeader && (x.SampleType==AudioSampleType.INT24 || x.SampleType==AudioSampleType.INT8))
+//										target.WriteByte(0);
 
 								}
 								if(m_process!=null)
@@ -276,8 +289,15 @@ namespace BeHappy
 		{
 			try
 			{
-				File.Delete(m_output);
-			}
+                if (m_bKeepOutput)
+                {
+                    raiseEvent(new EncoderCallbackEventArgs("Any output was kept for you to inspect."));
+                }
+                else
+                {
+                    File.Delete(m_output);
+                }
+            }
 			catch(Exception e2)
 			{
 				raiseEvent(new EncoderCallbackEventArgs(e2));
@@ -303,9 +323,9 @@ namespace BeHappy
 				// {3} means channel count
 				// {4} means samplecount
 				// {5} means size in bytes
-				info.Arguments = string.Format(m_commandLine, 
+				info.Arguments = string.Format(m_commandLine,
 					m_output, x.AudioSampleRate, x.BitsPerSample, x.ChannelsCount,m_nSampleCount,m_nSizeInBytes);
-				info.FileName = m_encoder;
+				info.FileName = m_encoder;				
 				raiseEvent(new EncoderCallbackEventArgs(string.Format("{0} {1}", m_encoder, info.Arguments)));
 				info.UseShellExecute = false;
 				info.RedirectStandardInput = true;
@@ -331,7 +351,7 @@ namespace BeHappy
 		private string saveScriptToTempFile()
 		{
 			string sTempFileName = System.IO.Path.GetTempPath() + "encode-" + Guid.NewGuid().ToString("n") + ".avs";
-	
+
 			using( Stream tempAVS = new FileStream(sTempFileName, System.IO.FileMode.CreateNew))
 			{
 				using( TextWriter avswr = new StreamWriter(tempAVS, System.Text.Encoding.Default))
@@ -351,12 +371,14 @@ namespace BeHappy
 			target.Write(BitConverter.GetBytes(useFaadTrick?FAAD_MAGIC_VALUE:(uint)(m_nSizeInBytes + WAV_HEADER_SIZE)),0,4);
 			target.Write(System.Text.Encoding.ASCII.GetBytes("WAVEfmt "),0,8);
 			target.Write(BitConverter.GetBytes((uint)0x10),0,4);
-#warning Format Tag
-			target.Write(BitConverter.GetBytes((short)0x01),0,2);
+//#warning Format Tag
+// New for float output, now is ok for allowed AudioSampleType. Pending FLOAT_64 and WAVE_FORMAT_EXTENSIBLE header.
+			target.Write(BitConverter.GetBytes((x.SampleType==AudioSampleType.FLOAT) ? ((short)0x03) : ((short)0x01)),0,2);
+// old			target.Write(BitConverter.GetBytes((short)0x01),0,2);
 			target.Write(BitConverter.GetBytes(x.ChannelsCount),0,2);
 			target.Write(BitConverter.GetBytes(x.AudioSampleRate),0,4);
 			target.Write(BitConverter.GetBytes(x.BitsPerSample*x.AudioSampleRate*x.ChannelsCount/8),0,4);
-			target.Write(BitConverter.GetBytes( x.ChannelsCount * x.BitsPerSample/8),0,2);
+			target.Write(BitConverter.GetBytes(x.ChannelsCount * x.BitsPerSample/8),0,2);
 			target.Write(BitConverter.GetBytes(x.BitsPerSample),0,2);
 			target.Write(System.Text.Encoding.ASCII.GetBytes("data"),0,4);
             target.Write(BitConverter.GetBytes(useFaadTrick ? (FAAD_MAGIC_VALUE - WAV_HEADER_SIZE) : (uint)m_nSizeInBytes), 0, 4);
@@ -392,6 +414,11 @@ namespace BeHappy
 				return m_encoderThread!=null && m_encoderThread.IsAlive;
 			}
 		}
+
+        internal void SetKeepOutput(bool bKeepOutput)
+        {
+            m_bKeepOutput = bKeepOutput;
+        }
 	}
 }
 /*
